+++ app-backend/application/src/contract_implementations/passage_query_impl.rs
@@ -18,7 +18,7 @@ impl PassageQueryImpl {
 }
 
 impl PassageQueries for PassageQueryImpl {
-    fn find_passage_between_locations(&self, from_location_id: i32, to_location_id: i32) -> BoxFuture<'static, Result<Option<Passage>, Error>> {
+    fn find_passage_between_locations(&self, from_location_id: u64, to_location_id: u64) -> BoxFuture<'static, Result<Option<Passage>, Error>> {
         let repo = self.passage_repository.clone();
         Box::pin(async move {
             let passage_dto_result = repo.find_by_start_and_end_id(from_location_id, to_location_id).await;
@@ -26,7 +26,7 @@ impl PassageQueries for PassageQueryImpl {
         })
     }
 
-    fn find_passage_by_location_and_direction(&self, location_id: i32, direction: &str) -> BoxFuture<'static, Result<Option<Passage>, Error>> {
+    fn find_passage_by_location_and_direction(&self, location_id: u64, direction: &str) -> BoxFuture<'static, Result<Option<Passage>, Error>> {
         let repo = self.passage_repository.clone();
         let direction = direction.to_owned();  // Clone to capture in async block
         Box::pin(async move {
