+++ app-backend/port/src/dto/passage_dto.rs
@@ -1,9 +1,9 @@
 use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct PassageDTO {
-    pub id: i32,
-    pub from_location_id: i32,
-    pub to_location_id: i32,
+    pub id: u64,
+    pub from_location_id: u64,
+    pub to_location_id: u64,
     pub description: String,
     pub direction: String,
     pub narration: String,
