+++ app-backend/port/src/dto/player_state_dto.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct PlayerStateDTO {
-    pub player_id: i32,
-    pub current_location_id: i32,
+    pub player_id: u64,
+    pub current_location_id: u64,
 }
